+++ packages/backend/src/server/api/stream/channels/global-timeline.ts
@@ -20,6 +20,7 @@ class GlobalTimelineChannel extends Channel {
 	public static requireCredential = false;
 	private withReplies: boolean;
 	private withFiles: boolean;
+	private withRenotes: boolean;
 
 	constructor(
 		private metaService: MetaService,
@@ -38,7 +39,8 @@ class GlobalTimelineChannel extends Channel {
 		const policies = await this.roleService.getUserPolicies(this.user ? this.user.id : null);
 		if (!policies.gtlAvailable) return;
 
-		this.withReplies = params.withReplies as boolean;
+		this.withReplies = params.withReplies ?? false;
+		this.withRenotes = params.withRenotes ?? true;
 		this.withFiles = params.withFiles as boolean;
 
 		// Subscribe events
@@ -72,6 +74,8 @@ class GlobalTimelineChannel extends Channel {
 			// 「チャンネル接続主への返信」でもなければ、「チャンネル接続主が行った返信」でもなければ、「投稿者の投稿者自身への返信」でもない場合
 			if (reply.userId !== this.user!.id && note.userId !== this.user!.id && reply.userId !== note.userId) return;
 		}
+
+		if (note.renote && note.text == null && (note.fileIds == null || note.fileIds.length === 0) && !this.withRenotes) return;
 
 		// Ignore notes from instances the user has muted
 		if (isInstanceMuted(note, new Set<string>(this.userProfile?.mutedInstances ?? []))) return;