+++ packages/frontend/src/store.ts
@@ -115,10 +115,6 @@ export const defaultStore = markRaw(new Storage('base', {
 		where: 'account',
 		default: [] as string[],
 	},
-	showTimelineReplies: {
-		where: 'account',
-		default: false,
-	},
 
 	menu: {
 		where: 'deviceAccount',