+++ packages/misskey-js/src/entities.ts
@@ -655,4 +655,7 @@ export type ModerationLog = {
 } | {
 	type: 'unmarkSensitiveDriveFile';
 	info: ModerationLogPayloads['unmarkSensitiveDriveFile'];
+} | {
+	type: 'createInvitation';
+	info: ModerationLogPayloads['createInvitation'];
 });