+++ CHANGELOG.md
@@ -12,6 +12,18 @@
 
 -->
 
+## next
+
+### General
+- Enhance: タイムラインからRenoteを除外するオプションを追加
+- Enhance: ユーザーページのノート一覧でRenoteを除外できるように
+
+### Client
+- Enhance: モデレーションログ機能の強化
+
+### Server
+- Enhance: MasterプロセスのPIDを書き出せるように
+
 ## 2023.9.1
 
 ### General